[Refactor] 추상 팩토리 패턴을 팩토리 메소드 패턴으로 변경
