+++ app/build.gradle
@@ -2,6 +2,9 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
+//apply plugin: 'com.google.firebase.crashlytics'
+//apply plugin: 'com.google.gms.google-services'
+//apply plugin: 'io.fabric'
 // deleted apply plugin: 'com.google.gms.google-services' due to build error
 
 android {
@@ -78,11 +81,11 @@ dependencies {
     implementation "com.squareup.retrofit2:adapter-rxjava2:2.6.2"
 
     //firebase
-    implementation 'com.google.firebase:firebase-core:17.2.3'
-    implementation 'com.google.firebase:firebase-analytics:17.2.3'
+    implementation 'com.google.firebase:firebase-core:17.4.0'
+    implementation 'com.google.firebase:firebase-analytics:17.4.0'
     // add SDKs for any other desired Firebase products
     // https://firebase.google.com/docs/android/setup#available-libraries
-    implementation 'com.google.firebase:firebase-analytics:17.3.0'
+    implementation 'com.google.firebase:firebase-crashlytics:17.0.0'
 
     // Naver Map
     implementation 'com.naver.maps:map-sdk:3.7.1'
