+++ app/src/main/java/com/golddog/mask_location/ext/MapExt.kt
@@ -6,12 +6,10 @@ import androidx.core.content.ContextCompat
 import com.golddog.mask_location.R
 import com.golddog.mask_location.entity.HospitalClinic
 import com.golddog.mask_location.entity.StoreSales
-import com.golddog.mask_location.ext.status.*
 import com.golddog.mask_location.ui.dialog.InfoBottomSheet
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.NaverMap
 import com.naver.maps.map.overlay.Marker
-import com.naver.maps.map.overlay.OverlayImage
 
 fun setMarkerVisible(markers: List<Marker>, naverMap: NaverMap) {
     markers.forEach {
@@ -56,19 +54,7 @@ fun setStoreMarker(
     var infoString = SpannableString("")
     marker.position = LatLng(storeSales.lat, storeSales.lng)
 
-    if (status == "plenty") {
-        infoString = StatusFactory.getStatus(PlentyStatusFactory(), context, marker, storeSales)
-    } else if (status == "some") {
-        infoString = StatusFactory.getStatus(SomeStatusFactory(), context, marker, storeSales)
-    } else if (status == "few") {
-        infoString = StatusFactory.getStatus(FewStatusFactory(), context, marker, storeSales)
-    } else if (status == "empty") {
-        infoString = StatusFactory.getStatus(EmptyStatusFactory(), context, marker, storeSales)
-    } else if (status == "break") {
-        infoString = StatusFactory.getStatus(BreakStatusFactory(), context, marker, storeSales)
-    } else {
-        marker.map = null
-    }
+    infoString = IntoStringFactory.createInfoString(status, context, marker, storeSales)
 
     marker.setOnClickListener {
         infoBottomSheet.showWithInfo(context, infoString)
