+++ build.gradle
@@ -5,14 +5,18 @@ buildscript {
     repositories {
         google()
         jcenter()
-        
+        maven {
+            url 'https://maven.fabric.io/public'
+        }
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.6.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.google.gms:google-services:4.3.3'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
+        classpath 'com.google.gms:google-services:4.3.3'
+//        classpath 'io.fabric.tools:gradle:1.31.2'
+//        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.0.0'
     }
 }
